Allow enabling timelapse generation (off by default)
